Add new library project NotesAPI_SharedModels + move shared models from NotesAPI to the new project + update FrontendMVC to reference new project + remove duplicated models from FrontendMVC project
