+++ FrontendMVC/FrontendMVC.csproj
@@ -16,6 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\NotesAPI_SharedModels\NotesAPI_SharedModels.csproj" />
     <ProjectReference Include="..\PatientsAPI_SharedModels\PatientsAPI_SharedModels.csproj" />
   </ItemGroup>
