--- FrontendMVC/Models/NoteModel.cs
@@ -1,13 +1,0 @@
-﻿namespace FrontendMVC.Models
-{
-    public class NoteModel
-    {
-        public string NoteId { get; set; }
-
-        public int PatientId { get; set; }
-
-        public string Content { get; set; }
-
-        public DateTime CreatedAt { get; set; }
-    }
-}
