--- FrontendMVC/Models/NoteModelAdd.cs
@@ -1,9 +1,0 @@
-﻿namespace FrontendMVC.Models
-{
-    public class NoteModelAdd
-    {
-        public int PatientId { get; set; }
-
-        public string Content { get; set; }
-    }
-}
