--- FrontendMVC/Models/NoteModelUpdate.cs
@@ -1,7 +1,0 @@
-﻿namespace FrontendMVC.Models
-{
-    public class NoteModelUpdate
-    {
-        public string Content { get; set; }
-    }
-}
