+++ FrontendMVC/Views/Patients/UpdateNote.cshtml
@@ -1,4 +1,4 @@
-﻿@model FrontendMVC.Models.NoteModelUpdate
+﻿@model NotesAPI_SharedModels.NoteModelUpdate
 
 <h2>Modifier une Note</h2>
