+++ MediLabo.sln
@@ -17,6 +17,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FrontendMVC", "FrontendMVC\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PatientsAPI_SharedModels", "PatientsAPI_SharedModels\PatientsAPI_SharedModels.csproj", "{501BACAC-C22E-4063-ACD2-D4F8DF826CA6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NotesAPI_SharedModels", "NotesAPI_SharedModels\NotesAPI_SharedModels.csproj", "{5025B3DF-1559-45D0-B08C-1A52660C10C9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -51,6 +53,10 @@ Global
 		{501BACAC-C22E-4063-ACD2-D4F8DF826CA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{501BACAC-C22E-4063-ACD2-D4F8DF826CA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{501BACAC-C22E-4063-ACD2-D4F8DF826CA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5025B3DF-1559-45D0-B08C-1A52660C10C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5025B3DF-1559-45D0-B08C-1A52660C10C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5025B3DF-1559-45D0-B08C-1A52660C10C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5025B3DF-1559-45D0-B08C-1A52660C10C9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
